Move Eventing.Tests Helpers tests into Interop and Readers
